Remove <ScrollView> or pass in scrollEnabled prop
Great component! : ) 
Nevertheless, I had to remove ScrollView in order for my form components to stay fixed. Otherwise, any minimal vertical movement would trigger the vertical scroll which is not something desirable, especially with a range select component. I would propose removing ScrollView or allowing the scrollEnabled prop to be passed in, so that it's possible to create fixed position forms.